refactor: reorganize folder structure for UI components and update imports
